Move tests to SheetContainer for access to toolbar and rangelist.
