+++ packages/plugins/plugin-sheet/src/components/GridSheet/GridSheet.stories.tsx
@@ -7,35 +7,16 @@ import '@dxos-theme';
 import { type Meta } from '@storybook/react';
 import React from 'react';
 
-import { IntentProvider, type Intent } from '@dxos/app-framework';
-import { useSpace, create } from '@dxos/react-client/echo';
+import { useSpace } from '@dxos/react-client/echo';
 import { withClientProvider } from '@dxos/react-client/testing';
 import { withTheme, withLayout } from '@dxos/storybook-utils';
 
 import { GridSheet } from './GridSheet';
 import { createTestCells, useTestSheet, withComputeGraphDecorator } from '../../testing';
 import translations from '../../translations';
-import { SheetAction, SheetType } from '../../types';
+import { SheetType } from '../../types';
 import { useComputeGraph } from '../ComputeGraph';
 import { SheetProvider } from '../SheetContext';
-
-// TODO(thure): Should we have a decorator for this?
-const storybookIntentValue = create({
-  dispatch: async (intents: Intent | Intent[]) => {
-    const intent = Array.isArray(intents) ? intents[0] : intents;
-    switch (intent.action) {
-      case SheetAction.DROP_AXIS: {
-        if (!intent.undo) {
-          const { model, axis, axisIndex } = intent.data as SheetAction.DropAxis;
-          model[axis === 'col' ? 'dropColumn' : 'dropRow'](axisIndex);
-        }
-      }
-    }
-  },
-  undo: async () => ({}),
-  history: [],
-  registerResolver: () => () => {},
-});
 
 export const Basic = () => {
   const space = useSpace();
@@ -46,28 +27,9 @@ export const Basic = () => {
   }
 
   return (
-    <IntentProvider value={storybookIntentValue}>
-      <SheetProvider graph={graph} sheet={sheet} ignoreAttention>
-        <GridSheet />
-      </SheetProvider>
-    </IntentProvider>
-  );
-};
-
-export const Spec = () => {
-  const space = useSpace();
-  const graph = useComputeGraph(space);
-  const sheet = useTestSheet(space, graph, { cells: { A1: { value: 'Ready' } } });
-  if (!sheet || !graph) {
-    return null;
-  }
-
-  return (
-    <IntentProvider value={storybookIntentValue}>
-      <SheetProvider graph={graph} sheet={sheet} ignoreAttention>
-        <GridSheet />
-      </SheetProvider>
-    </IntentProvider>
+    <SheetProvider graph={graph} sheet={sheet} ignoreAttention>
+      <GridSheet />
+    </SheetProvider>
   );
 };
